% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hglm.R
\name{as.OP.hglm_model}
\alias{as.OP.hglm_model}
\title{Convert to OP}
\usage{
\method{as.OP}{hglm_model}(x)
}
\arguments{
\item{x}{an object inheriting from \code{"hglm_model"}.}
}
\value{
A \pkg{ROI} object of class \code{"OP"}.
}
\description{
Convert an object of class \code{hglm_model} to \code{ROI::OP}.
}
