% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{cov_matrix}
\alias{cov_matrix}
\title{Construct Covariance matrix}
\usage{
cov_matrix(k, i, j, v)
}
\arguments{
\item{k}{an integer giving the number of rows and columns of the constructed 
covariance matrix.}

\item{i}{an integer vector giving the row indices.}

\item{j}{an integer vector giving the row indices.}

\item{v}{a numeric vector giving the corresponding values.}
}
\value{
A dense \code{matrix} of covariances.
}
\description{
Utility function for constructing covariance matrices based on 
a simple triplet format (\code{\link[slam]{simple_triplet_matrix}}).
}
\examples{
cov_matrix(5, c(1, 2), c(2, 3), c(0.8, 0.9))
}
