% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constraints.R
\name{hglmc}
\alias{hglmc}
\alias{c.hglmc}
\alias{is.hglmc}
\title{Generic Functions for \code{hglmc} Objects}
\usage{
\method{c}{hglmc}(...)

is.hglmc(x)
}
\arguments{
\item{...}{multiple objects inheriting from \code{"hglmc"} to be combined.}

\item{x}{an R object.}
}
\value{
An object of class \code{"hglmc"}.
}
\description{
Generic functions for holistic 'GLM' constraints.
}
