\docType{package}
\name{holiglm-package}
\alias{holiglm-package}
\title{holiglm: Holistic Generalized Linear Models}
\description{
The holistic generalized linear models package simplifies estimating GLMs
under constraints by making use of convex optimization.
}
\author{
\strong{Maintainer}: Benjamin Schwendinger \email{benjaminschwe@gmail.com}

Authors:
\itemize{
  \item Florian Schwendinger
  \item Laura Vana
}

}
