% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hglm.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.OP}
\alias{solution}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ROI}{\code{\link[ROI:OP]{as.OP}}, \code{\link[ROI]{solution}}}
}}

