% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi.R
\name{solution.hglm}
\alias{solution.hglm}
\title{Extract Solution}
\usage{
\method{solution}{hglm}(
  x,
  type = c("primal", "dual", "aux", "psd", "msg", "objval", "status", "status_code"),
  force = FALSE,
  ...
)
}
\arguments{
\item{x}{an object of type \code{'hglm'}.}

\item{type}{a character giving the name of the solution to be extracted.}

\item{force}{a logical to control the return value in the case that the
status code is equal to 1 (i.e. something went wrong).
By default force is \code{FALSE} and a solution is only provided
if the status code is equal to 0 (i.e. success). If force is \code{TRUE}
the status code is ignored and solutions are returned also
where the solver signaled an issue.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the extracted solution.
}
\description{
The solution of the underlying optimization problem,
 can be accessed via the method \code{'solution'}.
}
