% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hglm.R
\name{update_objective}
\alias{update_objective}
\title{Update the Model Object}
\usage{
update_objective(model, op)
}
\arguments{
\item{model}{an object inheriting from \code{"hglm_model"}.}

\item{op}{an \code{\link[ROI]{OP}} giving the new objective.}
}
\description{
Update the Model Object
}
